requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
pydantic